refactor(agent): sever the last type-only imports of supervisor-side modules
The agent side still imported aleph.vm.models and aleph.vm.hypervisors for
types only. Dissolve those edges so the agent's import closure no longer
executes any supervisor-side module (prerequisite hygiene for the supervisor
port):

- custom_logs: drop the dead ctx_current_execution ContextVar (it had no
  set-site anywhere, so the branch reading it could only ever see None) and
  with it the VmExecution import.
- run: create_vm_execution and its HTTP wrapper always return None per their
  docstrings and every return statement; annotate them as -> None and drop
  the VmExecution import.
- MigrationState moves from aleph.vm.models to aleph.vm.agent.migration.jobs,
  next to the ExportJob/ImportJob dataclasses that carry it (the agent is its
  only consumer).
- RuntimeConfiguration moves from aleph.vm.hypervisors.firecracker.microvm to
  aleph.vm.supervisor_interface.types: it is guest-init protocol vocabulary
  both sides interpret (this also frees the neutral aleph.vm.program_config
  from its hypervisors import).
- import-linter: add aleph.vm.models and aleph.vm.hypervisors to the agent
  contract's forbidden_modules; the residual-edges comment is gone because
  there are no residuals left.
