+++ tests/supervisor/test_local_migration_p2p.py
@@ -17,11 +17,11 @@ from aleph_message.models import ItemHash
 
 from aleph.vm.agent.migration.jobs import (
     ExportJob,
+    MigrationState,
     _reset_migration_semaphore_for_tests,
 )
 from aleph.vm.agent.migration.runner import run_export
 from aleph.vm.conf import settings
-from aleph.vm.models import MigrationState
 from aleph.vm.supervisor_interface.abc import Supervisor
